Fixed some bugs in parser, set up semanalyzer
I think a useful primitive for the analyzer will be a combination of a
fold and a map which I have brilliantly dubbed "mapfold"; it does a
combination of both, folding at the same time as mapping, so that the
mapping function has access to the folded value along the way. This
fits, for example, the behavior of the analyzer: it maps untyped terms
to typed terms, and carries the state of the analyzer along with
it (which state stores things like the symbol table).